Organize endpoints & eslint
