+++ backend/.eslintrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+  env: {
+    node: true,
+    commonjs: true,
+    es2021: true,
+  },
+  extends: "eslint:recommended",
+  parserOptions: {
+    ecmaVersion: "latest",
+  },
+  rules: {},
+};
