+++ backend/app.js
@@ -4,11 +4,10 @@ const express = require("express");
 const path = require("path");
 const cookieParser = require("cookie-parser");
 const logger = require("morgan");
-
-const indexRouter = require("./routes/index");
-const proposals = require("./routes/proposals");
+const indexRouter = require("./routes");
 const cors = require("cors");
 const app = express();
+
 app.use(cors());
 require("./utils/mongodb/connection");
 
@@ -18,7 +17,7 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
-app.use("/app", proposals);
+// routes
 app.use("/", indexRouter);
 
 // catch 404 and forward to error handler
@@ -27,7 +26,7 @@ app.use(function (req, res, next) {
 });
 
 // error handler
-app.use(function (err, req, res, next) {
+app.use(function (err, req, res) {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get("env") === "development" ? err : {};
