+++ backend/package.json
@@ -21,5 +21,8 @@
     "node-fetch": "^2.6.7",
     "node-html-markdown": "^1.1.3",
     "validator": "^13.7.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.10.0"
   }
 }
