+++ backend/routes/web.js
@@ -1,0 +1,12 @@
+var express = require("express");
+var router = express.Router();
+const path = require("path");
+/* GET home page. */
+router.get("/", function (req, res) {
+  res.sendFile(path.join(__dirname, "../public", "index.html"));
+});
+router.get("*", function (req, res) {
+  res.sendFile(path.join(__dirname, "../public", "index.html"));
+});
+
+module.exports = router;
