+++ frontend/src/pages/HomePage.svelte
@@ -8,7 +8,7 @@
   let projects;
 
   async function fetchProjects() {
-    const res = await fetch(`${SERVER_URI}/app/myProjects`, {
+    const res = await fetch(`${SERVER_URI}/app/project/list`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -21,7 +21,7 @@
     const data = await res.json();
     projects = data;
 
-    console.log(projects)
+    console.log(projects);
   }
   fetchProjects();
 
@@ -30,8 +30,27 @@
   }
 </script>
 
+<div class="flex h-screen home-container">
+  <Section
+    class="flex text-left bg-grey-200"
+    title={"DAO Projects"}
+    description={"Welcome to the OceanDAO Proposal Portal. Create projects, submit proposals, and complete them to access higher funding."}
+    descriptionTextLeft
+    actions={[
+      {
+        text: "Create Project",
+        onClick: onCreateNewProject,
+      },
+    ]}
+  >
+    {#if projects}
+      <ProjectItemsList {projects} />
+    {/if}
+  </Section>
+</div>
+
 <style>
-  .home-container{
+  .home-container {
     height: 100%;
     max-width: 800px;
     flex-direction: column;
@@ -39,18 +58,3 @@
     padding-top: var(--spacer);
   }
 </style>
-
-<div class="flex h-screen home-container">
-  <Section class="flex text-left bg-grey-200"
-    title={"DAO Projects"}
-    description={"Welcome to the OceanDAO Proposal Portal. Create projects, submit proposals, and complete them to access higher funding."}
-    descriptionTextLeft
-    actions={[{
-      "text": "Create Project",
-      "onClick":  onCreateNewProject
-    }]}>
-    {#if projects}
-      <ProjectItemsList {projects} />
-    {/if}
-  </Section>
-</div>
