+++ frontend/src/pages/ProposalPage.svelte
@@ -1,40 +1,42 @@
 <script>
-  import {SERVER_URI} from "../utils/config";
-  import moment from 'moment';
-  import {getNonce} from "../utils/helpers";
-  import {userAddress, networkSigner} from "../stores/ethers";
+  import { SERVER_URI } from "../utils/config";
+  import moment from "moment";
+  import { getNonce } from "../utils/helpers";
+  import { userAddress, networkSigner } from "../stores/ethers";
   import Swal from "sweetalert2";
-  import {signMessage} from "../utils/signatures";
-  import earmarks from '../utils/types/earmark.json'
+  import { signMessage } from "../utils/signatures";
+  import earmarks from "../utils/types/earmark.json";
   import Section from "../components/Section.svelte";
   import AdminEarmarkStatus from "../components/AdminEarmarkStatus.svelte";
   import AdminDeliverableStatus from "../components/AdminDeliverableStatus.svelte";
 
   export let proposalId;
   let proposal;
-  let deliverableActions = []
+  let deliverableActions = [];
 
   let pageText = {
     proposalDescription: `Use the form below to submit your final deliverables and complete your proposal. This enables your project to remain in a good state, and to apply for more grants.`,
-    submissionDescription: `Thank you for submitting your deliverables. Our admins are reviewing your update, and will provide an update soon.`
-  }
+    submissionDescription: `Thank you for submitting your deliverables. Our admins are reviewing your update, and will provide an update soon.`,
+  };
 
   async function loadData() {
-    let res = await fetch(`${SERVER_URI}/app/proposalInfo/${proposalId}`);
+    let res = await fetch(`${SERVER_URI}/app/proposal/info/${proposalId}`);
     proposal = await res.json();
 
-    if( proposal ) {
-      if( proposal.delivered.status === 0 || proposal.delivered.status === 2 ) {
-        deliverableActions = [{
-          "text": "Submit Deliverables",
-          "onClick": onSubmitDeliverableClick
-        }]
+    if (proposal) {
+      if (proposal.delivered.status === 0 || proposal.delivered.status === 2) {
+        deliverableActions = [
+          {
+            text: "Submit Deliverables",
+            onClick: onSubmitDeliverableClick,
+          },
+        ];
       }
     }
 
-    console.log(proposal)
-    console.log(proposal.proposalEarmark)
-    console.log(earmarks)
+    console.log(proposal);
+    console.log(proposal.proposalEarmark);
+    console.log(earmarks);
   }
   loadData();
 
@@ -70,9 +72,9 @@
         const json = await res.json();
         if (json.success === true) {
           Swal.fire(
-                  "Success!",
-                  "You've successfully withdrawn your proposal",
-                  "success"
+            "Success!",
+            "You've successfully withdrawn your proposal",
+            "success"
           ); // ? Popup flashes & goes away without user interaction. Looks broken. Proposal view does not render.
         } else {
           Swal.fire("Error!", "Something went wrong", "error");
@@ -90,60 +92,72 @@
   }
 </script>
 
+<!-- TODO - Visualize Admin: Is proposal completion submitted? Accepted? Rejected? -->
+<div class="flex h-screen flex-col justify-center proposal-container">
+  {#if proposal}
+    <Section
+      title={proposal.proposalTitle}
+      description={proposal.proposalDescription}
+      descriptionBottom
+      actions={[
+        {
+          text: "Withdraw Proposal",
+          onClick: withdrawProposal,
+        },
+        {
+          text: "Update Proposal",
+          onClick: onUpdateProposalClick,
+        },
+      ]}
+    >
+      <div class="details py-5 px-5">
+        <div class="col-start-4 col-span-2 ...">
+          <span class="detailName font-bold">Earmarks</span>
+          <span class="text-lg detailValue"
+            >{earmarks[proposal.proposalEarmark]}</span
+          >
+        </div>
+        <div class="col-start-7 col-span-2 ...">
+          <span class="detailName font-bold">Creation date</span>
+          <span class="text-lg detailValue"
+            >{moment(proposal.createdAt).format("YYYY-MM-DD")}</span
+          >
+        </div>
+      </div>
+      <AdminEarmarkStatus
+        proposalEarmarkRequest={proposal.proposalEarmarkRequest}
+      />
+    </Section>
+  {/if}
+  {#if proposal}
+    <Section
+      title={"Complete Proposal"}
+      description={pageText.proposalDescription}
+      actions={deliverableActions}
+    >
+      <AdminDeliverableStatus
+        deliverableStatus={proposal.delivered.status}
+        adminDescription={proposal.delivered.adminDescription}
+      />
+    </Section>
+  {/if}
+</div>
+
 <style>
-  .proposal-container{
+  .proposal-container {
     height: 100%;
     max-width: 800px;
     margin: auto;
     padding-top: var(--spacer);
   }
-  .details{
+  .details {
     display: flex;
     align-items: center;
     justify-content: space-between;
     background-color: var(--background-grey-dimmed);
   }
-  .detailName, .detailValue{
+  .detailName,
+  .detailValue {
     font-size: var(--font-size-normal);
   }
 </style>
-
-<!-- TODO - Visualize Admin: Is proposal completion submitted? Accepted? Rejected? -->
-<div class="flex h-screen flex-col justify-center proposal-container">
-  {#if proposal }
-    <Section
-            title={proposal.proposalTitle}
-            description={proposal.proposalDescription}
-            descriptionBottom
-            actions={[
-            {
-              "text": "Withdraw Proposal",
-              "onClick":  withdrawProposal
-            },{
-              "text": "Update Proposal",
-              "onClick":  onUpdateProposalClick
-            }]}>
-      <div class="details py-5 px-5">
-        <div class="col-start-4 col-span-2 ...">
-          <span class="detailName font-bold">Earmarks</span>
-          <span class="text-lg detailValue">{earmarks[proposal.proposalEarmark]}</span>
-        </div>
-        <div class="col-start-7 col-span-2 ...">
-          <span class="detailName font-bold">Creation date</span>
-          <span class="text-lg detailValue">{moment(proposal.createdAt).format('YYYY-MM-DD')}</span>
-        </div>
-      </div>
-      <AdminEarmarkStatus proposalEarmarkRequest={proposal.proposalEarmarkRequest}/>
-    </Section>
-  {/if}
-  {#if proposal}
-    <Section
-            title={"Complete Proposal"}
-            description={pageText.proposalDescription}
-            actions={deliverableActions}>
-      <AdminDeliverableStatus
-            deliverableStatus={proposal.delivered.status}
-            adminDescription={proposal.delivered.adminDescription} />
-    </Section>
-  {/if}
-</div>
